Introducing a new table in OVSDB to keep track of OVS-DPDK hw-acceleration features.
Updating the OVSDB schema with a new table to keep track of hardware acceleration
features that are relevant to OVS. Each PF/hardware device will populate a row
in this table with their properties.

The new hw table is introduced due to following reasons

1) Hardware acceleration in userspace datapath is handled completely via DPDK.
It is not necessary that control interface/tools are present in kernel for
querying or configuring underlying hardware. Since DPDK offers only APIs to
access these details, OVSDB is used to display it to the users/orchestrator.

2) Similar to DPDK parameters, hardware acceleration does have parameters such
as number of virtual ports to use in OVS, allocated memory, hw features to
enable and etc. OVSDB stores those information for orchestrator similar to the
how its been done for DPDK parameters.

Change-Id: I056c1a32e9a6ec84d155097e352d987696b43f38
Signed-off-by: Chandran, Sugesh <sugesh.chandran@intel.com>