Adding new netdev for hardware port with a hmap.
A new netdev structure is being defined for hardware accelerated ports.
Existing dpdk ports are not extended for enabling hardware acceleration due
to following reasons.

1) Easy to maintain and test. Since the hardware acceleration doesnt operate
on DPDK structures/functions, no need to worry about impact of these changes on
existing DPDK ports. It also avoid unnecessary condition check on DPDK port
functions to validate if port is accelerated or not.

2) Most of the DPDK structure elements are not relevant to hardware accelerated
ports and vice-versa. So keeping it separate make more sense than having it
in one structure.

In future, we may refactor the code in such a way to move common code between
hardware acceleration and normal DPDK ports into one place.

Each hardware accelerated port has a hmap to cache its hw flows. New
functions are added to operate on these hash-map. This cache optimize the flow
+ stat read from hardware by reducing the number of bytes to read from.

There is a possible optimization present in this commit by having a hmap per
hw-port than a single global hmap. This reduces the cost of lookup when there
are many flows scattered across the ports.

Change-Id: I0739b608b2be8ecc09a298f9543fcb88a23e167b
Signed-off-by: Chandran, Sugesh <sugesh.chandran@intel.com>