Backporting netdev flow APIs.
Kernel datapath flow offload is not available in OVS-2.7. So backporting
the netdev flow API definitions and their implementations.

DPDK hardware offload uses the same netdev APIs for the hw acceleration
as in kernel datapath. These changes are now available in the latest
OVS codebase. There is no new functionality introduced with this commit.

Change-Id: Ide39e6c06841fc1faaaf6191253cac9d92a2eff2
Signed-off-by: Chandran, Sugesh <sugesh.chandran@intel.com>