Update netdev hashmap on port add/delete.
When a new hardware accelerated port is added to a dpif-netdev(DPDK datapath),
it will be cached in a hash-map for operations on those ports in future.

An eg: of adding a hardware accelerated port via cli is given below.

 ovs-vsctl add-port br0 dpdkhw1 -- set Interface dpdkhw1 type=dpdkhwvhost options:device-id=0,port-id=1

The above eg: will add a hardware accelerated port '1' from device 0 into the OVS
bridge br0. The device-id is being populated based on the hardware devices that
are added at the time of vswitchd init. Each hardware device will get
a device-id and can be seen in OVSDB 'hw_offload' table.

In this eg:, qemu uses vhost socket '<hw-vhost-dir>/hw-vhost-user-1'
to connect to the 'dpdkhw1' port. The 'port-id'(1) is being used in vhost socket
name 'hw-vhost-user-1' to map to the right hardware vhost backend.

similarly for 'port-id=2' vhost socket name would be 'hw-vhost-user-2'
for 'port-id=3' vhost socket name would be 'hw-vhost-user-3' and so on.

Here is how its looks like in 'hw_offload' table when a device is initialized
successfully

hw_offload table
_uuid                                dev_id features                              name       pci_id
------------------------------------ ------ ------------------------------------- ---------- --------------
9b5a2336-e408-4425-a0fa-ca040663631a 0      {n_phy_ports="2", n_vhost_ports="32"} "switch-0" "0000:5e:00.0"

this 'dev_id' in the table is used later when adding ports to OVS.

Change-Id: I3b3b1e4f7e9fc48b1d04b7a4ddb4879d9d0d6adc
Signed-off-by: Chandran, Sugesh <sugesh.chandran@intel.com>